Refactor: Move Types into matching Files
