+++ MPSSELight/mpsse/Bit.cs
@@ -1,0 +1,8 @@
+namespace MPSSELight
+{
+    public enum Bit
+    {
+        One = 1,
+        Zero = 0,
+    }
+}
